+++ .eslintrc.json
@@ -26,6 +26,7 @@
         "react/jsx-filename-extension": "off",
         "react/jsx-one-expression-per-line": "off",
         "react/no-array-index-key": "off",
-        "arrow-parens": "off"
+        "arrow-parens": "off",
+        "react/jsx-props-no-spreading": "off"
     }
 }
