+++ src/components/App.js
@@ -1,13 +1,12 @@
 import React from 'react';
-import '../styles/App.scss';
 import Display from './Display';
-import ButtonPannel from './ButtonPannel';
+import ButtonPanel from './ButtonPanel';
 
 function App() {
   return (
-    <div className="App">
+    <div className="app">
       <Display value="0" />
-      <ButtonPannel />
+      <ButtonPanel />
     </div>
   );
 }
