+++ src/components/Display.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-const Display = ({ value }) => (<div>{value}</div>);
+const Display = ({ value }) => (<div className="display">{value}</div>);
 
 Display.propTypes = {
   value: PropTypes.string.isRequired,
